Revert "Devops"
