--- devops/nginx_basic.conf
@@ -1,53 +1,0 @@
-server {
-    listen 80;
-    server_name _;
-
-    location / {
-        include proxy_params;
-        proxy_pass http://127.0.0.1:5000;
-    }
-
-    location /static/ {
-        alias <path-to-your-application>/static/;
-        expires 30d;
-    }
-
-    location /socket.io {
-        include proxy_params;
-        proxy_http_version 1.1;
-        proxy_buffering off;
-        proxy_set_header Upgrade $http_upgrade;
-        proxy_set_header Connection "Upgrade";
-        proxy_pass http://127.0.0.1:5000/socket.io;
-    }
-}
-
-
-
-server {
-    listen 80;
-
-    location / {
-        proxy_pass http://127.0.0.1:5000;
-        proxy_http_version 1.1;
-        proxy_set_header Upgrade $http_upgrade;
-        proxy_set_header Connection "Upgrade";
-        proxy_set_header Host $host;
-
-          # Add CORS headers
-        #        add_header 'Access-Control-Allow-Origin' '*';
-        #add_header 'Access-Control-Allow-Methods' 'GET, POST, OPTIONS';
-        #add_header 'Access-Control-Allow-Headers' 'Origin, Content-Type, Accept, Authorization';
-        #add_header 'Access-Control-Allow-Credentials' 'true';
-
-        # Handle preflight requests (OPTIONS)
-        #if ($request_method = OPTIONS) {
-        #   add_header 'Access-Control-Allow-Origin' '*';
-        #   add_header 'Access-Control-Allow-Methods' 'GET, POST, OPTIONS';
-        #   add_header 'Access-Control-Allow-Headers' 'Origin, Content-Type, Accept, Authorization';
-        #   return 204; # No Content
-        #}
-    }
-}
-
-~
