--- devops/nginx_sites_enabled.conf
@@ -1,24 +1,0 @@
-server {
-    listen 80;
-    server_name api.haikuer.auguilin.com;
-    location / {
-        return 301 https://$host$request_uri;
-    }
-}
-
-# HTTPS server
-server {
-    listen 443 ssl;
-    server_name api.haikuer.auguilin.com;
-
-    ssl_certificate /etc/letsencrypt/live/api.haikuer.auguilin.com/fullchain.pem; # managed by Certbot
-    ssl_certificate_key /etc/letsencrypt/live/api.haikuer.auguilin.com/privkey.pem; # managed by Certbot
-
-    location / {
-        proxy_pass http://127.0.0.1:5000;
-        proxy_http_version 1.1;
-        proxy_set_header Upgrade $http_upgrade;
-        proxy_set_header Connection "Upgrade";
-        proxy_set_header Host $host;
-    }
-}
