--- devops/services/gunicorn.service
@@ -1,11 +1,0 @@
-[Unit]
-Description=Gunicorn Service for Haikuer App
-After=network.target
-
-[Service]
-ExecStart=gunicorn --worker-class eventlet -w 1 -b 0.0.0.0:5000 serve:app
-WorkingDirectory=/home/ubuntu/haikuer/backend/
-Restart=always
-
-[Install]
-WantedBy=multi-user.target
