+++ package.json
@@ -28,12 +28,9 @@
     "eslint-plugin-react-refresh": "^0.4.12",
     "gh-pages": "^6.2.0",
     "globals": "^15.9.0",
-    "notistack": "^3.0.1",
     "postcss": "^8.4.47",
     "postcss-preset-mantine": "^1.17.0",
     "postcss-simple-vars": "^7.0.1",
-    "react-router-dom": "^6.27.0",
-    "socket.io-client": "^4.8.0",
     "syllable": "^5.0.1",
     "vite": "^5.4.8"
   }
