+++ src/App.css
@@ -2,12 +2,10 @@
     margin-top: 5em;
     padding-top: 2em;
     padding-bottom: 2em;
-    margin-bottom: 20px;
     align-items: center;
     display: flex;
     flex-direction: column;
     background-color: #fff;
-    border-radius: 2em;
 }
 
 .buttons-class{
@@ -17,28 +15,4 @@
 button{
     margin-left: 5px;
     margin-right: 5px;
-}
-
-.container-class:focus-within {
-    border: 0.5px solid black;
-}
-
-.canvas {
-    width: 400px;
-    height: 300px;
-    overflow: scroll;
-    -ms-overflow-style: none; /* Internet Explorer 10+ */
-    scrollbar-width: none; /* Firefox */
-
-    /* Hide scrollbar for Chrome, Safari and Opera */
-    &::-webkit-scrollbar {
-        display: none;
-    }
-}
-
-/* Responsive for phones */
-@media (max-width: 600px) {
-    .canvas {
-        width: 100vw;
-    }
 }
