+++ src/main.jsx
@@ -2,7 +2,6 @@ import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import App from './App.jsx'
 import './index.css'
-import { SnackbarProvider } from 'notistack';
 
 // module.exports = {
 //   plugins: {
@@ -21,8 +20,6 @@ import { SnackbarProvider } from 'notistack';
 
 createRoot(document.getElementById('root')).render(
   <StrictMode>
-    <SnackbarProvider>
-      <App />
-    </SnackbarProvider>
+    <App />
   </StrictMode>,
 )
