--- src/multiplyaer/socketService.jsx
@@ -1,13 +1,0 @@
-import { io } from 'socket.io-client';
-
-const socket = io('https://api.haikuer.auguilin.com');
-
-export function initializeSocket() {
-  socket.connect();
-}
-
-export function disconnectSocket() {
-  if (socket) socket.disconnect();
-}
-
-export default socket;
