+++ vite.config.js
@@ -3,6 +3,6 @@ import react from '@vitejs/plugin-react'
 
 // https://vitejs.dev/config/
 export default defineConfig({
-  plugins: [react()]
-//  base: '/haikuer/'
+  plugins: [react()],
+  base: '/haikuer/'
 })
